+++ services/table-geocoder/lib/internal-geocoder/abstract_query_generator.rb
@@ -1,0 +1,35 @@
+# encoding: utf-8
+
+require_relative 'input_type_resolver'
+
+module CartoDB
+  module InternalGeocoder
+
+    class AbstractQueryGenerator
+
+      def initialize(internal_geocoder)
+        @internal_geocoder = internal_geocoder
+      end
+
+      #TODO custom exception
+      def search_terms_query(page)
+        raise 'Not implemented'
+      end
+
+      def dataservices_query(search_terms)
+        raise 'Not implemented'
+      end
+
+      def copy_results_to_table_query
+        raise 'Not implemented'
+      end
+
+      def country
+        country = @internal_geocoder.countries
+        country == %Q{'world'} ? 'null' : country
+      end
+
+    end # AbstractQueryGenerator
+
+  end # InternalGeocoder
+end #CartoDB