+++ NEWS.md
@@ -1,11 +1,21 @@
-3.2.3 (2014-XX-XX)
+3.2.4 (2014-XX-XX)
 ------------------
 * ArcGIS import integration
 
-3.2.2 (2014-09-26)
+3.2.3 (2014-09-30)
+------------------
+* New features
+  * Added ability to disable high resolution geocoding batch api by config 
+* Fixed bugs
+  * Missing geocoding type (Admin regions, country column, polygons)
+
+3.2.2 (2014-09-29)
 ------------------
 * New features
   * Improved torque rendering on pause or static datasets
+  * Improved mobile layout
+  * Improved Geocoder UI
+  * Refactor of InternalGeocoder
 
 3.2.1 (2014-09-05)
 ------------------